+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 3.6.2
 
 - Fixed: XML Comment on parameter had too many `>` characters, i.e. `// <param name="param">>Test Parameter</param>` 
+- Fixed: CRUD Mapping issue around constructors with ValueObjects as parameters. 
 
 ### Version 3.6.1